Before datetime change.

+++ src/actions/advertisement-types/deleteAdvertisementType.ts
@@ -5,7 +5,7 @@ import { auth } from "@/auth";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
-const deleteAdvertisementType = async (formData: FormData) => {
+const deleteAdvertisementType = async (advertisementId: string) => {
   try {
     const session = await auth();
     if (!session) {
@@ -17,15 +17,37 @@ const deleteAdvertisementType = async (formData: FormData) => {
         },
       };
     }
-    const advertisementId = formData.get("advertisementId")?.toString() || "-1";
+    
     const userId = session.user?.id;
-
-    await prisma.advertisement.delete({
+    const advertisement = await prisma.advertisement.findFirst({
       where: {
         id: advertisementId,
         userId,
       },
+      select: {
+        id: true,
+        purchases: true,
+      },
     });
+
+    // Check if purchases exist for the advertisement and if so soft delete the advertisement
+    if (advertisement?.purchases && advertisement?.purchases.length > 0) {
+      await prisma.advertisement.update({
+        where: {
+          id: advertisementId,
+        },
+        data: {
+          isDeleted: true,
+        },
+      });
+    } else {
+      await prisma.advertisement.delete({
+        where: {
+          id: advertisementId,
+          userId,
+        },
+      });
+    }
   } catch (error: any) {
     // Handle any potential errors here
     console.error("Error deleting advertisement type", error);
