+++ src/actions/calendar-editions/deleteCalendarEdition.ts
@@ -4,7 +4,7 @@ import prisma from "@/lib/prisma/prisma";
 import { auth } from "@/auth";
 import { revalidatePath } from "next/cache";
 
-const deleteCalendar = async (formData: FormData) => {
+const deleteCalendar = async (calendarId: string) => {
   try {
     const session = await auth();
     if (!session) {
@@ -16,16 +16,36 @@ const deleteCalendar = async (formData: FormData) => {
         },
       };
     }
-    const calendarId = formData.get("calendarId")?.toString() || "-1";
     const userId = session.user?.id;
 
     await prisma.$transaction(async (prisma) => {
-      await prisma.calendarEdition.delete({
+      const calendar = await prisma.calendarEdition.findFirst({
         where: {
           id: calendarId,
           userId,
         },
+        select: {
+          id: true,
+          purchases: true,
+        },
       });
+      if (calendar?.purchases && calendar?.purchases.length > 0) {
+        await prisma.calendarEdition.update({
+          where: {
+            id: calendarId,
+          },
+          data: {
+            isDeleted: true,
+          },
+        });
+      } else {
+        await prisma.calendarEdition.delete({
+          where: {
+            id: calendarId,
+            userId,
+          },
+        });
+      }
     });
 
   } catch (error: any) {
