+++ src/actions/contact/deleteContact.ts
@@ -4,7 +4,7 @@ import prisma from "@/lib/prisma/prisma";
 import { auth } from "@/auth";
 import { revalidatePath } from "next/cache";
 
-const deleteContact = async (formData: FormData) => {
+const deleteContact = async (contactId: string) => {
   try {
     const session = await auth();
     if (!session) {
@@ -16,20 +16,40 @@ const deleteContact = async (formData: FormData) => {
         },
       };
     }
-    const contactId = formData.get("contactId")?.toString() || "-1";
     const userId = session.user?.id;
 
     await prisma.$transaction(async (prisma) => {
-      await prisma.contactAddressBook.deleteMany({
-        where: {
-          contactId: contactId,
-        },
-      });
-  
-      await prisma.contact.delete({
+      const contact = await prisma.contact.findFirst({
         where: {
           id: contactId,
           userId,
+        },
+        select: {
+          id: true,
+          payments: true,
+        },
+      });
+      if (contact?.payments && contact?.payments.length > 0) {
+        await prisma.contact.update({
+          where: {
+            id: contactId,
+          },
+          data: {
+            isDeleted: true,
+          },
+        });
+      } else {
+        await prisma.contact.delete({
+          where: {
+            id: contactId,
+            userId,
+          },
+        });
+
+      }
+      await prisma.contactAddressBook.deleteMany({
+        where: {
+          contactId: contactId,
         },
       });
     })
